[BOTF-259] Home Page Content (#257)
* [BOTF-260] Binance API Section (#250)
Created the Binance API section on the Home Page

* [BOTF-261] Choose a Bot Section (#251)
Created the Choose a Bot section

* [BOTF-262] Build a Bot Section
Created the Build Bot section

* [BOTF-263] Adjust small issues (#253)
Adjusted several issues that found on the home page.

* [BOTF-264] Add transition animation on scrolling for the rest of elements (#254)
Added all transitions needed

* [BOTF-265] Issue on email confirmation (#255)
Fixed the issue

* [BOTF-266] New view for Dummy Rule (#256)
Created a second view for the dummy rule tile

* [BOTF-267] Issues with the main page wrap on login page (#258)
The login page is a no-scroll page. that was the issue, it was getting scroll.

* [BOTF-268] Issues with the error alert message (#259)
Improved and fixed the error alert modal

* [BOTF-270] But on alert message topbar position (#260)
Fixed the issue by creating a portal.

* [BOTF-271] Fix the images alert messages on the console (#261)
Added the fill property to fix the issue.